Fix tour item hover background in navigation
Not all list items in the navigation are of the same height. If any item
other than "workflow features" was active there would be a gap between
the a-element's background color and end of the navigation list item.

By making the a-element have a height of 100% we fill the entire
navigation list item, filling the gap.