Split text utilities
In order to keep the files that contain utilities small and searchable,
I split out the inline text to the categories it belongs to.

Since we added a `text-size` file, I took the time to also add all the
other text sizes so we are feature complete on that one.

Text utilities is something we would use pretty often, so I also took
the libery to shorten some of the utilities, and made the block selector
into `.u-t`.

To make them easier to guess when writing css I removed the mapping to
HTML, and named them after what they do.