Add @responsive helper to generate responsive CSS
When creating utilities you sometimes want to have them generated in a
couple of variants for different screens. To prevent we have to do all
of this by hand I created a @responsive helper that you can use to wrap
the css in. All the classes inside that block will then get the
proper breakpoints generated. Keep in mind that the original block will
also always remain present.