Convert utilities to non-bem
It doesn't make sense to give utilities a modifier since there is
nothing to modify.

To give an example:

When you have a `.c-box`, you can have a modifier of `--with-spacing` to
indicate that there is a variation on the `c-box` that has extra
spacing.

In the case of utilities this doesn't work, when you take the `.u-t`
class for example, there is no base for that, only the actual utilities
like `u-t-underline`.

To make this more clear we are removing the `--` from the utility
classes.

Keep in mind that we are keeping the BEM style modifiers for all the
other css that we create, just not for utilitlies