Remove Lint from Build step
We have this in CI now, so there is no longer a reason to have this as
part of the Build step.

Main reason for removing this is that it won't adhere to the ignore file
that is setup for stylelint. So it keeps warning about violations in
dependencies in `node_modules`.